% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_standata_event.R
\name{make_standata_event}
\alias{make_standata_event}
\title{Create Stan data input for the event submodel}
\usage{
make_standata_event(
  formulaEvent,
  dataEvent,
  id_var,
  standata_long,
  assoc,
  a_K = 4,
  basehaz = "bs",
  basehaz_aux = list(df = 6, knots = NULL, degree = 3),
  qnodes = 15L
)
}
\arguments{
\item{formulaEvent}{Formula for the time-to-event submodel (without association terms)}

\item{dataEvent}{Event dataset}

\item{id_var}{Character string for name of ID variable in \code{dataEvent}.}

\item{standata_long}{An object returned by \code{make_standata_long.R}}

\item{assoc}{Type of association (either 'RE' for random effects, 'LP' for linear predictor, 'CV' for current value).}

\item{a_K}{Number of association terms (use 4: mean and WIV for each biomarker)}

\item{basehaz}{Type of baseline hazard function (only 'bs' is implemented)}

\item{basehaz_aux}{List of parameters for the baseline hazard function.}

\item{qnodes}{Number of quadrature nodes.}

\item{M}{Number of longitudinal biomarkers (at the moment, only 2)}

\item{time_varLong}{Character string for name of time variable in \code{dataLong}.}
}
\value{
A list to be passed to \code{make_standata_jmwiv.R}.
}
\description{
Create Stan data input for the event submodel
}
