% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_standata_jmwiv.R
\name{make_standata_jmwiv}
\alias{make_standata_jmwiv}
\title{Create Stan data input object for JM-WIV}
\usage{
make_standata_jmwiv(
  formulaLong1,
  formulaLong2 = NULL,
  dataLong,
  id_var,
  time_varLong,
  formulaEvent,
  dataEvent,
  assoc,
  a_K = 4,
  basehaz = "bs",
  basehaz_aux = list(df = 6, knots = NULL, degree = 3),
  qnodes = 15L,
  prior_list = NULL
)
}
\arguments{
\item{formulaLong1}{Formula for the first longitudinal marker.}

\item{formulaLong2}{Formula for the second longitudinal marker.}

\item{dataLong}{Longitudinal dataset.}

\item{id_var}{Character string for name of ID variable in \code{dataLong}.}

\item{time_varLong}{Character string for name of time variable in \code{dataLong}.}

\item{formulaEvent}{Formula for the time-to-event submodel (without association terms)}

\item{dataEvent}{Event dataset}

\item{assoc}{Type of association (either 'RE' for random effects, 'LP' for linear predictor, 'CV' for current value).}

\item{a_K}{Number of association terms}

\item{basehaz}{Type of baseline hazard function (only 'bs' is implemented)}

\item{basehaz_aux}{List of parameters for the baseline hazard function.}

\item{qnodes}{Number of quadrature nodes.}

\item{prior_list}{List of prior arguments.}
}
\value{
A list to be passed to \code{rstan::sampling}.
}
\description{
Create Stan data input object for JM-WIV
}
\examples{
library(rstanarm)
f_logBili <- list(formula(logBili ~ sex + trt + year + (1 | id)),
formula(sigma ~ sex + trt + year + (1 | id)))
f_albumin <- list(formula(albumin ~ sex + trt + year + (1 | id)),
formula(sigma ~ sex + trt + year + (1 | id)))
f_Event <- survival::Surv(futimeYears, death) ~ sex + trt
prior_input <- list(b_prior_scale = rep(1, 4),
                    b_prior_df = rep(1, 4),
                    b_prior_regularization = 5)
standata_jmwiv <- make_standata_jmwiv(formulaLong1 = f_logBili,
                         formulaLong2 = f_albumin,
                         dataLong = pbcLong,
                         id_var = "id",
                         time_varLong = "year",
                         formulaEvent = ,
                         dataEvent = pbcSurv,
                         assoc = "CV",
                         a_K = 4,
                         basehaz = "bs",
                         basehaz_aux = list(df = 6,
                                            knots = NULL,
                                            degree = 3),
                         qnodes = 15L,
                         prior_list = prior_input)
}
