% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_standata_long.R
\name{make_standata_long}
\alias{make_standata_long}
\title{Create Stan data input for the longitudinal (MELS) submodel}
\usage{
make_standata_long(
  formulaLong1,
  formulaLong2 = NULL,
  dataLong,
  id_var,
  time_varLong
)
}
\arguments{
\item{formulaLong1}{Formula for the first longitudinal marker.}

\item{formulaLong2}{Formula for the second longitudinal marker.}

\item{dataLong}{Longitudinal dataset.}

\item{id_var}{Character string for name of ID variable in \code{dataLong}.}

\item{time_varLong}{Character string for name of time variable in \code{dataLong}.}
}
\value{
A list to be passed to \code{make_standata_jmwiv.R}.
}
\description{
Create Stan data input for the longitudinal (MELS) submodel
}
\examples{
library(rstanarm)
data(pbcLong)
f_logBili <- list(formula(logBili ~ sex + trt + year + (1 | id)),
formula(sigma ~ sex + trt + year + (1 | id)))
f_albumin <- list(formula(albumin ~ sex + trt + year + (1 | id)),
formula(sigma ~ sex + trt + year + (1 | id)))
u1 <- make_standata_long(formulaLong1 = f_logBili,
                         formulaLong2 = f_albumin,
                         dataLong = pbcLong,
                         id_var = "id",
                         time_varLong = "year")
}
