% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_standata_prior.R
\name{make_standata_prior}
\alias{make_standata_prior}
\title{Create Stan data input for the prior distributions}
\usage{
make_standata_prior(standata_long, standata_event, prior_input = NULL)
}
\arguments{
\item{standata_long}{An output from \code{make_standata_long.R}.}

\item{standata_event}{An output from \code{make_standata_event.R}.}

\item{prior_input}{A list with prior parameters.}
}
\value{
A list with all prior parameters to be passed to \code{make_standata_jmwiv.R}.
}
\description{
Create Stan data input for the prior distributions
}
