% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jmwiv_functions.R
\name{ppcheck}
\alias{ppcheck}
\title{Posterior predictive checks for the longitudinal submodel of JM-WIV}
\usage{
ppcheck(list_of_draws, data, x, idvar)
}
\arguments{
\item{list_of_draws}{Output from \code{rstan::extract(fit)}.}

\item{data}{Longitudinal dataset.}

\item{x}{Time variable in the longitudinal dataset.}

\item{idvar}{ID variable in the longitudinal dataset.}
}
\value{
A dataframe with posterior predictive draws for each longitudinal biomarker.
}
\description{
Posterior predictive checks for the longitudinal submodel of JM-WIV
}
\examples{
ppcheck(list_of_draws = rstan::extract(fit_jmwiv), data = pbcLong,
                  x = "year", idvar = "id")
}
