% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jmwiv_functions.R
\name{pred_plot_jmwiv}
\alias{pred_plot_jmwiv}
\title{Posterior prediction from JM-WIV for an individual}
\usage{
pred_plot_jmwiv(
  stanfit,
  standata,
  id_selected,
  start_timezero = FALSE,
  var_labels,
  xlim_input = attr(standata$basehaz_X, "Boundary.knots"),
  plot_jm = TRUE
)
}
\arguments{
\item{stanfit}{Fit from \code{jmwiv_stan.R}}

\item{standata}{Stan data input}

\item{id_selected}{ID for one or more subject (if more than one value, you need to sort them)}

\item{start_timezero}{Start prediction at time 0 or at last longitudinal observation?}

\item{var_labels}{Labels for outcomes in the plot}

\item{xlim_input}{Range of x-axis}

\item{plot_jm}{Plot the predictions?}
}
\value{
Plot of prediction for one or more subjects.
}
\description{
Produce posterior predictions for one or more subjects (code is slow, please use only for up to 3 subjects).
}
\examples{
pred_plot_jmwiv(stanfit = fit_jmwiv,
standata = standata_jmwiv_pbc,
var_labels = c("logBilirubin", "Albumin", "Survival probability"),
id_selected = c(1,5), plot_jm = FALSE) +
 labs(x = "Age")
}
