% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jmwiv_functions.R
\name{prepare_standata_prediction}
\alias{prepare_standata_prediction}
\title{Extract covariates for submodel}
\usage{
prepare_standata_prediction(standata, param, times, ind)
}
\arguments{
\item{standata}{Stan data input}

\item{param}{Parameter}

\item{times}{Quadrature times}

\item{ind}{Index}
}
\value{
A dataframe to be used within \code{predict_long_qpts}.
}
\description{
Internal function
}
