% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jmwiv_functions.R
\name{ranef_corr_jmwiv}
\alias{ranef_corr_jmwiv}
\title{Print random effect correlation matrix}
\usage{
ranef_corr_jmwiv(fit)
}
\arguments{
\item{fit}{An object of class \code{stanfit} returned by \code{rstan::sampling}}
}
\value{
A correlation matrix.
}
\description{
Print random effect correlation matrix
}
