% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jmwiv_functions.R
\name{simulate_jmwiv}
\alias{simulate_jmwiv}
\title{Simulate data from a joint model from Cox specification with baseline hazard following exponential or Gompertz distribution and MELS specification for longitudinal submodel}
\usage{
simulate_jmwiv(
  seed,
  nsub,
  assoc = "RE",
  lambda = 1,
  g_shape = 0,
  e_beta = c(-0.5, 0),
  a_beta = c(-0.5, -0.2, -0.4, 0.3),
  distObs = 0.5,
  max_time = 10,
  ranef_covmat = c(3, 1, 2, 0.2) * diag(4),
  y1mu_fixed = c(-0.8, 2),
  y1sigma_fixed = c(0.2, -0.5),
  y2mu_fixed = c(1.2, -0.3),
  y2sigma_fixed = c(-0.6, 1.4)
)
}
\arguments{
\item{seed}{seed used for simulation}

\item{nsub}{number of subjects}

\item{assoc}{either random effects (RE) or linear predictor (LP) - current value association not available}

\item{lambda}{baseline hazard function (constant over time)}

\item{g_shape}{shape parameter for the Gompertz distribution (for g_shape=0, this reduces to Cox-exponential)}

\item{e_beta}{regression coefficient in Cox model for a baseline, binary variable (e.g. treatment)}

\item{a_beta}{regression coefficients in Cox model for the random intercept of the mean and SD of y1 and mean and SD of y2, respectively}

\item{distObs}{time between consecutive visits}

\item{max_time}{censoring time}

\item{ranef_covmat}{covariance matrix for random effects}

\item{y1mu_fixed}{fixed effects for y1_mu}

\item{y1sigma_fixed}{fixed effects for y1_sigma}

\item{y2mu_fixed}{fixed effects for y2_mu}

\item{y2sigma_fixed}{fixed effects for y2_sigma}
}
\value{
A list with a simulated dataset for the longitudinal data and a dataset for the event data.
}
\description{
This function returns a list with longitudinal and survival data.
The baseline hazard function is assumed to be constant over time.
At the moment only Cox-Exponential for 'RE' is available, while for 'LP' you can have Cox-Exp or Cox-Gompertz.
}
